Add files via upload
a. Month: The number of months based on user input in the “Number of Years” field
b. Opening Amount: The initial investment amount as well as the opening balance each
month, which includes interest
c. Deposited Amount: The dollar amount the user plans to deposit each month. This value
will be the same every month for the duration of the investment.
d. Total: The sum of the opening and deposited amounts
e. Interest: Money earned based on the “annual interest” rate input by the user. The
interest based on an opening amount of $1 and a deposited amount of $50 with an
interest rate of 5% compounded monthly is:
(Opening Amount + Deposited Amount) * ((Interest Rate/100)/12)
OR
(1 + 50) * ((5/100)/12)
Note: Dividing by 100 converts the interest rate percentage to a decimal.
Note: 12 is the number of months in a year. Dividing the yearly amount by twelve gives
monthly compounded interest.
f. Closing Balance: The sum of the total and interest amounts